app/requirements.txt
streamlit==1.38.0
ollama==0.3.3
groq==0.11.0
openai==1.52.0
langchain==0.3.2
langchain-core==0.3.9
langchain-groq==0.2.0
langchain_chroma==0.1.4
langchain-huggingface==0.1.0
langgraph==0.2.34
python-dotenv==1.0.1
mysql-connector-python==9.0.0
megaparse==0.0.31
llama-index-embeddings-langchain==0.2.1
llama-index-llms-groq==0.2.0
poppler-utils==0.1.0
pytesseract==0.3.13
openpyxl==3.1.5
pydantic==2.9.2
PyPDF2==3.0.1
PyMuPDF==1.24.13
anthropic==0.39.0